updating `dev` cache `logs` + `sitemap` update
