+++ src/routes/api/cache/seo_block/+server.ts
@@ -32,15 +32,9 @@ export async function GET(req: { url: { [x: string]: { get: (arg0: string) => st
 
 async function getCacheSEOblock (lang: string): Promise < Cache_Single_Homepage_SEO_Block_Translation_Response | Record < string, never > > {
   try {
-    // [ℹ] cached data retrival;
     const cached: string = await redis.hget('seo_block_t', lang);
-    // [ℹ] check for `cached` data
     if (cached) {
-      // [ℹ] convert the data from `string` to `JSON`
       const parsed: Cache_Single_Homepage_SEO_Block_Translation_Response = JSON.parse(cached);
-      // [🐛] debug;
-      if (dev) console.info("✅ seo_block_t cache HIT", lang);
-      // [ℹ] return, cached data;
       return parsed;
     }
     return
